The regex stuff moved from network.rs to regex_manager.rs
