+++ src/regex_manager.rs
@@ -2,9 +2,15 @@
 //! the [`crate::Engine`], infrequently used regexes can be discarded. The [`RegexManager`] is
 //! responsible for managing the storage of regexes used by filters.
 
-use crate::filters::network::{compile_regex, CompiledRegex, NetworkFilter};
+use crate::filters::network::NetworkFilter;
+
+use regex::{
+    bytes::Regex as BytesRegex, bytes::RegexBuilder as BytesRegexBuilder,
+    bytes::RegexSet as BytesRegexSet, bytes::RegexSetBuilder as BytesRegexSetBuilder, Regex,
+};
 
 use std::collections::HashMap;
+use std::fmt;
 use std::time::Duration;
 
 #[cfg(test)]
@@ -56,6 +62,40 @@ pub struct RegexDebugEntry {
     pub last_used: Instant,
     /// How many times this regex has been used.
     pub usage_count: usize,
+}
+
+#[derive(Debug, Clone)]
+pub enum CompiledRegex {
+    Compiled(BytesRegex),
+    CompiledSet(BytesRegexSet),
+    MatchAll,
+    RegexParsingError(regex::Error),
+}
+
+impl CompiledRegex {
+    pub fn is_match(&self, pattern: &str) -> bool {
+        match &self {
+            CompiledRegex::MatchAll => true, // simple case for matching everything, e.g. for empty filter
+            CompiledRegex::RegexParsingError(_e) => false, // no match if regex didn't even compile
+            CompiledRegex::Compiled(r) => r.is_match(pattern.as_bytes()),
+            CompiledRegex::CompiledSet(r) => {
+                // let matches: Vec<_> = r.matches(pattern).into_iter().collect();
+                // println!("Matching {} against RegexSet: {:?}", pattern, matches);
+                r.is_match(pattern.as_bytes())
+            }
+        }
+    }
+}
+
+impl fmt::Display for CompiledRegex {
+    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+        match &self {
+            CompiledRegex::MatchAll => write!(f, ".*"), // simple case for matching everything, e.g. for empty filter
+            CompiledRegex::RegexParsingError(_e) => write!(f, "ERROR"), // no match if regex didn't even compile
+            CompiledRegex::Compiled(r) => write!(f, "{}", r.as_str()),
+            CompiledRegex::CompiledSet(r) => write!(f, "{}", r.patterns().join(" | ")),
+        }
+    }
 }
 
 struct RegexEntry {
@@ -110,11 +150,87 @@ impl Default for RegexManager {
 
 fn make_regexp(filter: &NetworkFilter) -> CompiledRegex {
     compile_regex(
-        &filter.filter,
+        filter.filter.iter(),
         filter.is_right_anchor(),
         filter.is_left_anchor(),
         filter.is_complete_regex(),
     )
+}
+
+/// Compiles a filter pattern to a regex. This is only performed *lazily* for
+/// filters containing at least a * or ^ symbol. Because Regexes are expansive,
+/// we try to convert some patterns to plain filters.
+#[allow(clippy::trivial_regex)]
+pub(crate) fn compile_regex<'a, I>(
+    filters: I,
+    is_right_anchor: bool,
+    is_left_anchor: bool,
+    is_complete_regex: bool,
+) -> CompiledRegex
+where
+    I: Iterator<Item = &'a str> + ExactSizeIterator,
+{
+    use once_cell::sync::Lazy;
+    // Escape special regex characters: |.$+?{}()[]\
+    static SPECIAL_RE: Lazy<Regex> =
+        Lazy::new(|| Regex::new(r"([\|\.\$\+\?\{\}\(\)\[\]])").unwrap());
+    // * can match anything
+    static WILDCARD_RE: Lazy<Regex> = Lazy::new(|| Regex::new(r"\*").unwrap());
+    // ^ can match any separator or the end of the pattern
+    static ANCHOR_RE: Lazy<Regex> = Lazy::new(|| Regex::new(r"\^(.)").unwrap());
+    // ^ can match any separator or the end of the pattern
+    static ANCHOR_RE_EOL: Lazy<Regex> = Lazy::new(|| Regex::new(r"\^$").unwrap());
+
+    let mut escaped_patterns = Vec::with_capacity(filters.len());
+    for filter_str in filters {
+        // If any filter is empty, the entire set matches anything
+        if filter_str.is_empty() {
+            return CompiledRegex::MatchAll;
+        }
+        if is_complete_regex {
+            // unescape unrecognised escaping sequences, otherwise a normal regex
+            let unescaped = filter_str[1..filter_str.len() - 1]
+                .replace("\\/", "/")
+                .replace("\\:", ":");
+
+            escaped_patterns.push(unescaped);
+        } else {
+            let repl = SPECIAL_RE.replace_all(&filter_str, "\\$1");
+            let repl = WILDCARD_RE.replace_all(&repl, ".*");
+            // in adblock rules, '^' is a separator.
+            // The separator character is anything but a letter, a digit, or one of the following: _ - . %
+            let repl = ANCHOR_RE.replace_all(&repl, "(?:[^\\w\\d\\._%-])$1");
+            let repl = ANCHOR_RE_EOL.replace_all(&repl, "(?:[^\\w\\d\\._%-]|$)");
+
+            // Should match start or end of url
+            let left_anchor = if is_left_anchor { "^" } else { "" };
+            let right_anchor = if is_right_anchor { "$" } else { "" };
+            let filter = format!("{}{}{}", left_anchor, repl, right_anchor);
+
+            escaped_patterns.push(filter);
+        }
+    }
+
+    if escaped_patterns.is_empty() {
+        CompiledRegex::MatchAll
+    } else if escaped_patterns.len() == 1 {
+        let pattern = &escaped_patterns[0];
+        match BytesRegexBuilder::new(pattern).unicode(false).build() {
+            Ok(compiled) => CompiledRegex::Compiled(compiled),
+            Err(e) => {
+                // println!("Regex parsing failed ({:?})", e);
+                CompiledRegex::RegexParsingError(e)
+            }
+        }
+    } else {
+        match BytesRegexSetBuilder::new(escaped_patterns)
+            .unicode(false)
+            .build()
+        {
+            Ok(compiled) => CompiledRegex::CompiledSet(compiled),
+            Err(e) => CompiledRegex::RegexParsingError(e),
+        }
+    }
 }
 
 impl RegexManager {
